+++ src/test/java/org/mmarini/railways2/model/routes/DoubleSlipSwitchTest.java
@@ -42,7 +42,6 @@ import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.empty;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mmarini.railways.Matchers.emptyOptional;
-import static org.mmarini.railways.Matchers.optionalOf;
 
 class DoubleSlipSwitchTest {