+++ CHANGELOG.md
@@ -24,3 +24,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue #18: Change section discovery to handle crossing points
 - Issue #23: Add map panel
 - Issue #25: Add station panel
+- Issue #29: Add random arrival train